fix: Integer to Long

+++ src/main/java/jact/lagaltproject/exceptions/FreelancerNotFoundException.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(value = HttpStatus.NOT_FOUND)
 public class FreelancerNotFoundException extends RuntimeException {
-    public FreelancerNotFoundException(int id) {
+    public FreelancerNotFoundException(Long id) {
         super("Freelancer does not exist with that ID: " + id);
     }
 }
