+++ src/main/java/jact/lagaltproject/repositories/ProjectRepository.java
@@ -3,5 +3,5 @@ package jact.lagaltproject.repositories;
 import jact.lagaltproject.models.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectRepository extends JpaRepository<Project, Integer> {
+public interface ProjectRepository extends JpaRepository<Project, Long> {
 }
